requirements.txt
djangorestframework==3.12.4
dj-rest-auth
django-cors-headers>=3.14.0